updateProfile
เพิ่มที่เก็บ Profile

+++ database/src/application/useCases/closefriend/updateSatusFriend.js
@@ -10,10 +10,19 @@ class UpdateFriend {
             if (!['accepted', 'refuse'].includes(status)) {
                 throw new Error("Invalid status");
             }
+            const existingRequest = await this.closefriendService.checkrequest(UserID, friendid);
             if (status === 'accepted'){
-                await this.closefriendService.updatefriend(UserID, friendid, status);
+                if (existingRequest?.Status === 'accepted'){
+                    throw new Error("We are already friends.");
+                }else{
+                    await this.closefriendService.updatefriend(UserID, friendid, status);
+                }
             }else{
-                await this.closefriendService.deletefriend(UserID, friendid);
+                if (!existingRequest){
+                    throw new Error("This friend is not in the system.");
+                }else{
+                    await this.closefriendService.deletefriend(UserID, friendid);
+                }
             }
         } catch (error) {
             throw error;
